requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
passlib==1.7.4
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-multipart==0.0.9
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.5
